Tests/LibWeb/Text/expected/Fetch/fetch-later-no-activate-after.txt

activated before iframe unload: false
cookie observed before iframe unload: false
cookie observed after iframe unload: true
PASS
